+++ common/src/main/java/io/github/flemmli97/runecraftory/mixin/LevelRendererMixin.java
@@ -1,21 +1,31 @@
 package io.github.flemmli97.runecraftory.mixin;
 
+import com.mojang.blaze3d.vertex.PoseStack;
+import com.mojang.blaze3d.vertex.VertexConsumer;
+import io.github.flemmli97.runecraftory.client.ClientCalls;
 import io.github.flemmli97.runecraftory.common.utils.WorldUtils;
 import net.minecraft.client.Minecraft;
+import net.minecraft.client.multiplayer.ClientLevel;
 import net.minecraft.client.renderer.LevelRenderer;
 import net.minecraft.core.BlockPos;
+import net.minecraft.world.entity.Entity;
 import net.minecraft.world.level.biome.Biome;
+import net.minecraft.world.level.block.state.BlockState;
 import org.spongepowered.asm.mixin.Mixin;
 import org.spongepowered.asm.mixin.Shadow;
 import org.spongepowered.asm.mixin.Unique;
 import org.spongepowered.asm.mixin.injection.At;
+import org.spongepowered.asm.mixin.injection.Inject;
 import org.spongepowered.asm.mixin.injection.ModifyVariable;
+import org.spongepowered.asm.mixin.injection.callback.CallbackInfo;
 
 @Mixin(LevelRenderer.class)
 public abstract class LevelRendererMixin {
 
     @Shadow
     private Minecraft minecraft;
+    @Shadow
+    private ClientLevel level;
 
     @Unique
     private Biome rf4CacheBiome;
@@ -44,5 +54,10 @@ public abstract class LevelRendererMixin {
         if (WorldUtils.coldEnoughForSnow(this.minecraft.level, pos, this.rf4CacheBiome))
             return new BlockPos(pos.getX(), this.minecraft.level.getMinBuildHeight() - 1, pos.getZ());
         return pos;
+    }
+
+    @Inject(method = "renderHitOutline", at = @At("RETURN"))
+    private void onBlockOutline(PoseStack poseStack, VertexConsumer consumer, Entity entity, double camX, double camY, double camZ, BlockPos pos, BlockState state, CallbackInfo info) {
+        ClientCalls.onBlockHighlightRender(this.level, poseStack, consumer, entity, camX, camY, camZ, pos, state);
     }
 }