+++ Changelog.md
@@ -10,6 +10,7 @@ RuneCraftory 1.4.4
 - Make npc reset POI when too far away or can't reach
 - Make mobs that hold weapons able to use them
 - Make skill leveling when hitting entities tag based instead of item based
+- All fertilizer apply in a 2x2 now
 
 RuneCraftory 1.4.3
 ================